MAINT move maxtimemanager to shared util
